refactor(prover): address review findings on the memory PR
Trace row and stage 8:
- TraceRow::new returns Result<_, TraceRowError>; TraceError gains
  InvalidRow. Deserialize goes through the same constructor (TraceRowWire).
- Split trace.rs into trace/row.rs and sparse.rs into sparse/{layout,ops}.rs.
- OpeningView keys on PCS::OPENING_FOLD_IS_TERMINAL instead of assuming Dory
  folds once; the forwarded MultilinearPoly surface is documented.
- stage_boundary is gated on PURGE_MIN_LOG_T like every other purge site;
  release_retained_memory is private, purge_staging renamed
  purge_retained_memory.

RAM address column:
- Back to Vec<u32> with the typed encode check; the u64 widening kept 4*T
  extra bytes live across stages 2-6b with no 64-bit consumer.
- SharedRamAddresses derives Allocative natively so the Arc dedups; the
  custom visitor counted the column twice in stage-4/5 heap snapshots.
- collect_full documents the two-pass trade (u32 addresses outlive values;
  stage 4 re-reads the trace instead of parking 16*T bytes).

One owner per rule:
- Raw-column bind schedule (bound_pair, bind_raw_twice, mul_0_optimized)
  lives in support.rs; ram_val_check and the register sparse path call it.
- The 8x shrink law moves into Polynomial::bind_low_to_high_in_place, which
  reports the release; Polynomial::{capacity, shrink_to_fit} are gone.
- MatrixEntry::bind / type Bound become inherent methods; the trait keeps
  only the generic accumulate_pair_evals.
- TraceRow::is_noop (test-only caller) removed; bind_low_to_high_in_place
  free fn is private; row.instruction() is decoded once per row in the
  witness hot paths.

Tests: restore rejects_session_carry_from_another_cycle_domain,
rejects_register_outside_protocol_domain and
cycle_entry_count_matches_cycle_entries, all on still-live checks.

+++ crates/jolt-kernels/src/optimized/ram_ra_claim_reduction.rs
@@ -121,13 +121,13 @@ impl<F: JoltField> PrepareKernel<F, RamRaClaimReduction<F>> for OptimizedBackend
 /// `Q_x[c_lo] = Σ_{c_hi} eq(r_address)[addresses[c_hi‖c_lo]] · eq_hi_x[c_hi]`
 /// for the three cycle points, in one pass over the access columns.
 fn build_q_tables<F: JoltField>(
-    addresses: &[u64],
+    addresses: &[u32],
     eq_address: &[F],
     eq_hi: &[Vec<F>; TERMS],
     prefix_bits: usize,
 ) -> [Vec<F>; TERMS] {
     let prefix_size = 1usize << prefix_bits;
-    let fill = |q: &mut [Vec<F>; TERMS], base: usize, chunk: &[u64]| {
+    let fill = |q: &mut [Vec<F>; TERMS], base: usize, chunk: &[u32]| {
         for (i, &address) in chunk.iter().enumerate() {
             if address == NO_ACCESS {
                 continue;
@@ -179,7 +179,7 @@ fn build_q_tables<F: JoltField>(
 /// — the partial evaluation of the address-folded `ra` at the prefix
 /// challenges, regathered from the access columns.
 fn gather_h_prime<F: JoltField>(
-    addresses: &[u64],
+    addresses: &[u32],
     eq_address: &[F],
     eq_prefix: &[F],
     prefix_bits: usize,
@@ -187,7 +187,7 @@ fn gather_h_prime<F: JoltField>(
 ) -> Vec<F> {
     let prefix_size = 1usize << prefix_bits;
     let suffix_size = 1usize << suffix_bits;
-    let fill = |h: &mut Vec<F>, base: usize, chunk: &[u64]| {
+    let fill = |h: &mut Vec<F>, base: usize, chunk: &[u32]| {
         for (i, &address) in chunk.iter().enumerate() {
             if address == NO_ACCESS {
                 continue;
@@ -248,7 +248,7 @@ enum Phase<F: JoltField> {
         q: [Vec<F>; TERMS],
         #[cfg_attr(feature = "allocative", allocative(visit = jolt_poly::visit_scalar_rows))]
         eq_hi: [Vec<F>; TERMS],
-        addresses: Arc<Vec<u64>>,
+        addresses: Arc<Vec<u32>>,
         #[cfg_attr(feature = "allocative", allocative(visit = jolt_poly::visit_scalars))]
         eq_address: Vec<F>,
         #[cfg_attr(feature = "allocative", allocative(visit = jolt_poly::visit_scalar_rows))]
